requirements.txt
selenium
requests
python-dotenv
pytest
pydantic
pydantic[email]
pandas